+++ packages/sit-onyx/src/components/OnyxStepper/types.ts
@@ -12,10 +12,6 @@ export type OnyxStepperProps = FormInjectedProps &
   Omit<OnyxFormElementProps, "modelValue" | "errorMessages" | "withCounter" | "maxlength"> &
   AutofocusProp & {
     /**
-     * Current value of the input.
-     */
-    modelValue?: number;
-    /**
      * Same as the native `name` attribute of `<input>`.
      * Used to reference the input in JavaScript or in submitted form data.
      */
@@ -35,7 +31,27 @@ export type OnyxStepperProps = FormInjectedProps &
     /**
      * Incremental step.
      */
-    step?: number;
+    /**
+     * Incremental step.
+     * @deprecated
+     */
+    step?: number; // step-mismatch + step-increment
+
+    /**
+     * The smallest allowed value and rounded precision
+     */
+    precision?: number; // step-mismatch => uses :step="props.precision" for the validation
+    /**
+     * The increment number
+     * @default precision is the default stepSize
+     */
+    stepSize?: number; //  step-increment => number which is used for increment/decrement
+
+    /**
+     * Ensure no wrong number can be inputed
+     */
+    stripStep?: boolean;
+
     /**
      * Specify how to provide automated assistance in filling out the input.
      * Some autocomplete values might required specific browser permissions to be allowed by the user.